+++ packages/react-native-editor/CHANGELOG.md
@@ -11,6 +11,10 @@ For each user feature we should also add a importance categorization label  to i
 
 ## Unreleased
 
+## 1.58.1
+-   [*] Global styles: Check for undefined values and merge user colors [#33707]
+-   [*] [Embed block] Disable paragraph transform [#33745]
+
 ## 1.58.0
 -   [***] New Block: Embed block. [#33452]